0.0.8/maintenance: introduce Yard and adjust console module
