+++ .yardopts
@@ -1,0 +1,1 @@
+lib/**/*.rb - README.rdoc CHANGELOG.md rdoc/*.rdoc
