+++ lib/sf_cli/console.rb
@@ -1,0 +1,5 @@
+require 'sf_cli/console/commands'
+
+include SfCli::Console::Commands
+
+use ARGV[0]
